Combine `opencast-header` and `navigation` into `header`
