+++ src/ui/header.js
@@ -2,7 +2,7 @@
 /** @jsx jsx */
 import { jsx } from 'theme-ui';
 
-import { NavLink, useLocation } from 'react-router-dom';
+import { Link, NavLink, useLocation } from 'react-router-dom';
 import { Fragment, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
@@ -14,6 +14,57 @@ import {
   faVideo,
 } from "@fortawesome/free-solid-svg-icons";
 
+
+// The header, including a logo on the left and the navigation on the right.
+export default function Header() {
+  return (
+    <header
+      sx={{
+        height: theme => theme.heights.headerHeight,
+        lineHeight: theme => theme.heights.headerHeight,
+        color: 'background',
+        display: 'flex',
+        justifyContent: 'space-between',
+        position: 'relative',
+        zIndex: 3,
+      }}
+    >
+      {/* This div is used just for the background color. We can't set it for
+          the parent element, as the navigation overlay would otherwise occlude
+          this background color */}
+      <div sx={{
+        backgroundColor: 'gray.0',
+        position: 'absolute',
+        zIndex: -3,
+        height: '100%',
+        width: '100%',
+      }}></div>
+      <Brand />
+
+      <Navigation />
+    </header>
+  );
+}
+
+const Brand = () => {
+  const location = useLocation();
+
+  return (
+    <Link to={{ pathname: "/", search: location.search }}>
+      <picture>
+        <source
+          media="(min-width: 920px)"
+          srcSet={`${process.env.PUBLIC_URL}/opencast-studio.svg`}
+        />
+        <img
+          src={`${process.env.PUBLIC_URL}/opencast-studio-small.svg`}
+          alt="Opencast Studio"
+          sx={{ height: 50 }}
+        />
+      </picture>
+    </Link>
+  );
+}
 
 // One element (link) in the navigation.
 const NavElement = ({ target, children, icon, ...rest }) => {
@@ -148,5 +199,3 @@ const Navigation = props => {
     </Fragment>
   );
 };
-
-export default Navigation;
